+++ packages/manager/cypress/e2e/core/stackscripts/create-stackscripts.spec.ts
@@ -20,7 +20,7 @@ import { cleanUp } from 'support/util/cleanup';
 import { createTestLinode } from 'support/util/linodes';
 import { interceptGetAllImages } from 'support/intercepts/images';
 import type { Image } from '@linode/api-v4';
-import { getFilteredImagesForImageSelect } from 'src/components/ImageSelectv2/utilities';
+import { getFilteredImagesForImageSelect } from 'src/components/ImageSelect/utilities';
 
 // StackScript fixture paths.
 const stackscriptBasicPath = 'stackscripts/stackscript-basic.sh';
