+++ packages/manager/src/components/ImageSelect/ImageOption.test.tsx
@@ -3,14 +3,14 @@ import React from 'react';
 import { imageFactory } from 'src/factories';
 import { renderWithTheme } from 'src/utilities/testHelpers';
 
-import { ImageOptionv2 } from './ImageOptionv2';
+import { ImageOption } from './ImageOption';
 
-describe('ImageOptionv2', () => {
+describe('ImageOption', () => {
   it('renders the image label', () => {
     const image = imageFactory.build({ eol: null });
 
     const { getByText } = renderWithTheme(
-      <ImageOptionv2 image={image} isSelected={false} listItemProps={{}} />
+      <ImageOption image={image} isSelected={false} listItemProps={{}} />
     );
 
     expect(getByText(image.label)).toBeVisible();
@@ -19,7 +19,7 @@ describe('ImageOptionv2', () => {
     const image = imageFactory.build();
 
     const { getByTestId } = renderWithTheme(
-      <ImageOptionv2 image={image} isSelected={false} listItemProps={{}} />
+      <ImageOption image={image} isSelected={false} listItemProps={{}} />
     );
 
     expect(getByTestId('os-icon')).toBeVisible();
@@ -28,7 +28,7 @@ describe('ImageOptionv2', () => {
     const image = imageFactory.build({ capabilities: ['cloud-init'] });
 
     const { getByLabelText } = renderWithTheme(
-      <ImageOptionv2 image={image} isSelected={false} listItemProps={{}} />,
+      <ImageOption image={image} isSelected={false} listItemProps={{}} />,
       { flags: { metadata: true } }
     );
 
@@ -40,7 +40,7 @@ describe('ImageOptionv2', () => {
     const image = imageFactory.build({ capabilities: ['distributed-sites'] });
 
     const { getByLabelText } = renderWithTheme(
-      <ImageOptionv2 image={image} isSelected={false} listItemProps={{}} />
+      <ImageOption image={image} isSelected={false} listItemProps={{}} />
     );
 
     expect(
@@ -54,7 +54,7 @@ describe('ImageOptionv2', () => {
     const image = imageFactory.build({ deprecated: true });
 
     const { getByText } = renderWithTheme(
-      <ImageOptionv2 image={image} isSelected={false} listItemProps={{}} />
+      <ImageOption image={image} isSelected={false} listItemProps={{}} />
     );
 
     expect(getByText(`${image.label} (deprecated)`)).toBeVisible();
@@ -67,7 +67,7 @@ describe('ImageOptionv2', () => {
     });
 
     const { getByText } = renderWithTheme(
-      <ImageOptionv2 image={image} isSelected={false} listItemProps={{}} />
+      <ImageOption image={image} isSelected={false} listItemProps={{}} />
     );
 
     expect(getByText(`${image.label} (deprecated)`)).toBeVisible();
