+++ packages/manager/src/components/ImageSelect/utilities.test.ts
