+++ packages/manager/src/features/Linodes/LinodeCreate/Tabs/Images.tsx
@@ -5,8 +5,8 @@ import { useController, useFormContext, useWatch } from 'react-hook-form';
 import DistributedRegionIcon from 'src/assets/icons/entityIcons/distributed-region.svg';
 import ImageIcon from 'src/assets/icons/entityIcons/image.svg';
 import { Box } from 'src/components/Box';
-import { ImageSelectv2 } from 'src/components/ImageSelectv2/ImageSelectv2';
-import { getAPIFilterForImageSelect } from 'src/components/ImageSelectv2/utilities';
+import { ImageSelect } from 'src/components/ImageSelect/ImageSelect';
+import { getAPIFilterForImageSelect } from 'src/components/ImageSelect/utilities';
 import { Link } from 'src/components/Link';
 import { Paper } from 'src/components/Paper';
 import { Placeholder } from 'src/components/Placeholder/Placeholder';
@@ -101,7 +101,7 @@ export const Images = () => {
       <Paper>
         <Typography variant="h2">Choose an Image</Typography>
         <Box alignItems="flex-end" display="flex" flexWrap="wrap" gap={2}>
-          <ImageSelectv2
+          <ImageSelect
             disabled={isCreateLinodeRestricted}
             errorText={fieldState.error?.message}
             onBlur={field.onBlur}
